% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyprcoloc.R
\name{align.ABF.1}
\alias{align.ABF.1}
\title{align.ABF.1}
\usage{
align.ABF.1(
  Z,
  W,
  trait.cor,
  sample.overlap,
  ld.matrix,
  epsilon,
  reg.res,
  align.thresh,
  prior.1,
  prior.2,
  cor.adj.priors,
  uniform.priors,
  Zsq,
  Wsq,
  ind.traits
)
}
\arguments{
\item{Z}{matrix of Z-scores}

\item{W}{ratio matrix of prior standard deviation and observed standard errors}

\item{trait.cor}{correlation matrix between traits}

\item{sample.overlap}{matrix of sample overlap between traits}

\item{ld.matrix}{LD matrix}

\item{epsilon}{tolerance parameter}

\item{reg.res}{regional result}

\item{align.thresh}{alignment probability threshold}

\item{prior.1}{prior probability of a SNP being associated with one trait}

\item{prior.2}{1 - prior probability of a SNP being associated with an additional trait given that the SNP is associated with at least 1 other trait}

\item{cor.adj.priors}{correlation adjusted priors}

\item{uniform.priors}{uniform priors}

\item{Zsq}{matrix of Z-scores squared}

\item{Wsq}{matrix of W squared}

\item{ind.traits}{are the traits independent or to be treated as independent}
}
\description{
align.ABF.1
}
