% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyprcoloc.R
\name{align.ABF.2}
\alias{align.ABF.2}
\title{align.ABF.2}
\usage{
align.ABF.2(
  Z,
  W,
  snps.clc,
  trait.cor,
  sample.overlap,
  ld.matrix,
  epsilon,
  reg.res,
  align.thresh,
  prior.1,
  prior.2,
  prior.3,
  prior.4,
  cor.adj.priors,
  uniform.priors
)
}
\arguments{
\item{Z}{matrix of Z-scores}

\item{W}{ratio matrix of prior standard deviation and observed standard errors}

\item{snps.clc}{SNPs colocalisation}

\item{trait.cor}{correlation matrix between traits}

\item{sample.overlap}{matrix of sample overlap between traits}

\item{ld.matrix}{LD matrix}

\item{epsilon}{tolerance parameter}

\item{reg.res}{regional result}

\item{align.thresh}{alignment probability threshold}

\item{prior.1}{prior probability of a SNP being associated with one trait}

\item{prior.2}{1 - prior probability of a SNP being associated with an additional trait given that the SNP is associated with at least 1 other trait}

\item{prior.3}{prior probability that a trait contains a second causal variant given it contains one already}

\item{prior.4}{1 - prior probability that trait two co-localises with trait one given traits one and two already share a causal variant and trait one contains a second causal variant}

\item{cor.adj.priors}{correlation adjusted priors}

\item{uniform.priors}{uniform priors}
}
\description{
align.ABF.2
}
