% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyprcoloc.R
\name{cred.sets}
\alias{cred.sets}
\title{Credible sets}
\usage{
cred.sets(res, value = 0.95)
}
\arguments{
\item{res}{list of results from hyprcoloc when setting "snpscores = TRUE"}

\item{value}{the sum of the probabilities of the snps included in the credible set}
}
\description{
\code{cred.sets} computes credible sets of SNPs for each cluster of colocalized traits
}
\author{
Christopher Foley \href{mailto:chris.neal.foley@gmail.com}{chris.neal.foley@gmail.com} & James Staley \href{mailto:jrstaley95@gmail.com}{jrstaley95@gmail.com}
}
