% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyprcoloc.R
\name{hyprcoloc}
\alias{hyprcoloc}
\title{HyPrColoc}
\usage{
hyprcoloc(
  effect.est,
  effect.se,
  binary.outcomes = rep(0, dim(effect.est)[2]),
  trait.subset = c(1:dim(effect.est)[2]),
  trait.names = c(1:dim(effect.est)[2]),
  snp.id = c(1:dim(effect.est)[1]),
  ld.matrix = diag(1, dim(effect.est)[1], dim(effect.est)[1]),
  trait.cor = diag(1, dim(effect.est)[2], dim(effect.est)[2]),
  sample.overlap = matrix(rep(1, dim(effect.est)[2]^2), nrow = dim(effect.est)[2]),
  bb.alg = TRUE,
  bb.selection = "regional",
  reg.steps = 1,
  reg.thresh = "default",
  align.thresh = "default",
  prior.1 = 1e-04,
  prior.c = 0.02,
  prior.12 = NULL,
  sensitivity = FALSE,
  sense.1 = 1,
  sense.2 = 2,
  uniform.priors = FALSE,
  ind.traits = FALSE,
  snpscores = FALSE
)
}
\arguments{
\item{effect.est}{matrix of beta values}

\item{effect.se}{matrix of se values}

\item{binary.outcomes}{a binary vector depicting binary traits}

\item{trait.subset}{vector of traits from the full trait list for trageted coloclaisation analysis}

\item{trait.names}{vector of trait names corresponding to the columns in the effect.est matrix}

\item{snp.id}{vector of SNP IDs}

\item{ld.matrix}{LD matrix}

\item{trait.cor}{correlation matrix between traits}

\item{sample.overlap}{matrix of sample overlap between traits}

\item{bb.alg}{branch and bound algorithm}

\item{bb.selection}{branch and bound algorithm type}

\item{reg.steps}{regional step paramter}

\item{align.thresh}{alignment probability threshold}

\item{prior.1}{prior probability of a SNP being associated with one trait}

\item{prior.c}{conditional colocalization prior: prior probability of a SNP being associated with an additional trait given that the SNP is associated with at least 1 other trait}

\item{prior.12}{COLOC users: prior probability that a SNP is associated with two traits}

\item{sensitivity}{perform senstivity analysis}

\item{ind.traits}{are the traits independent or to be treated as independent}

\item{snpscores}{output estimated posterior probability explained each SNP}

\item{n.cvs}{number of causal variants}

\item{window.size}{size of window for 2CV testing}

\item{sentinel}{sentinel variant}

\item{epsilon}{tolerance parameter}

\item{reg.thres}{regional probability threshold}

\item{reg.tol}{regional tolerance parameter}

\item{prior.3}{prior probability that a trait contains a second causal variant given it contains one already}

\item{prior.4}{1 - prior probability that trait two co-localises with trait one given traits one and two already share a causal variant and trait one contains a second causal variant}

\item{sens.1}{first sensitivity analysis}

\item{sens.2}{second sensitivity analysis}

\item{cor.adj.priors}{correlation adjusted priors}

\item{unifrom.priors}{uniform priors}

\item{branch.jump}{branch jump}
}
\value{
a data.frame of HyPrColoc results

snpscores a list of estimated posterior probabilities explained by the SNPs; for the BB algorithm there is a set of SNP probabilities for each iteration
}
\description{
hyprcoloc is a function which allows the user to perform multi-trait colocalisation analyses in genomic regions
}
\examples{
# Regression coefficients and standard errors from ten GWAS studies (Traits 1-5, 6-8 & 9-10 colocalize)
betas <- hyprcoloc::test.betas
head(betas)
ses <- hyprcoloc::test.ses
head(ses)
  
# Trait names and SNP IDs
traits <- paste0("T", 1:10)
rsid <- rownames(betas)

# Colocalisation analyses
results <- hyprcoloc(betas, ses, trait.names=traits, snp.id=rsid)
}
\author{
Christopher Foley <christopher.foley@mrc-bsu.cam.ac.uk> and James R Staley <james.staley@bristol.ac.uk>
}
