% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyprmtc.R
\name{hyprmtc}
\alias{hyprmtc}
\title{HyPrMTC}
\usage{
hyprmtc(effect.est, effect.se, binary.outcomes = rep(0, dim(Z)[2]),
  trait.subset = c(1:dim(effect.est)[2]),
  trait.names = c(1:dim(effect.est)[2]), snp.id = c(1:dim(effect.est)[1]),
  ld.matrix = diag(1, dim(effect.est)[1], dim(effect.est)[1]),
  trait.cor = diag(1, dim(effect.est)[2], dim(effect.est)[2]),
  sample.overlap = matrix(rep(1, dim(effect.est)[2]^2), nrow =
  dim(effect.est)[2]), n.cvs = 1, bb.alg = TRUE, bb.type = "fast",
  test.2 = FALSE, reg.steps = 1, window.size = dim(effect.est)[1],
  sentinel = 0, epsilon = 0, reg.thresh = "default",
  align.thresh = "default", reg.tol = 0.699, prior.1 = 1e-04,
  prior.2 = 0.98, prior.3 = 0.001, prior.4 = 0.995, sensitivity = F,
  sense.1 = 1, sense.2 = 2, cor.adj.priors = FALSE,
  uniform.priors = TRUE, branch.jump = FALSE, ind.traits = TRUE)
}
\arguments{
\item{effect.est}{matrix of beta values}

\item{effect.se}{matrix of se values}

\item{binary.outcomes}{a binary vector depicting binary traits}

\item{trait.subset}{vector of traits from the full trait list for trageted coloclaisation analysis}

\item{trait.names}{vector of trait names corresponding to the columns in the effect.est matrix}

\item{snp.id}{vector of SNP IDs}

\item{ld.matrix}{LD matrix}

\item{trait.cor}{correlation matrix between traits}

\item{sample.overlap}{matrix of sample overlap between traits}

\item{n.cvs}{number of causal variants}

\item{bb.alg}{branch and bound algorithm}

\item{bb.type}{branch and bound algorithm type}

\item{reg.steps}{regional step paramter}

\item{window.size}{size of window for 2CV testing}

\item{sentinel}{sentinel variant}

\item{epsilon}{tolerance parameter}

\item{align.thresh}{alignment probability threshold}

\item{reg.tol}{regional tolerance parameter}

\item{prior.1}{prior probability of a SNP being associated with one trait}

\item{prior.2}{1 - prior probability of a SNP being associated with an additional trait given that the SNP is associated with at least 1 other trait}

\item{prior.3}{prior probability that a trait contains a second causal variant given it contains one already}

\item{prior.4}{1 - prior probability that trait two co-localises with trait one given traits one and two already share a causal variant and trait one contains a second causal variant}

\item{sensitivity}{perform senstivity analysis}

\item{cor.adj.priors}{correlation adjusted priors}

\item{branch.jump}{branch jump}

\item{ind.traits}{are the traits independent or to be treated as independent}

\item{reg.thres}{regional probability threshold}

\item{sens.1}{first sensitivity analysis}

\item{sens.2}{second sensitivity analysis}

\item{unifrom.priors}{uniform priors}
}
\value{
results data.frame of results

snp.scores list of snp scores for each set of colocalised traits
}
\description{
hyprmtc is a function which allows the user to perform multi-trait colocalisation analyses in genomic regions
}
\examples{
# Regression coefficients and standard errors from three GWAS studies
beta <- matrix(rnorm(3000), nrow=1000, ncol=3)
se <- matrix(abs(rnorm(3000)), nrow=1000, ncol=3)

# Colocalisation analyses
results <- hyprmtc(beta,se)
}
\author{
Christopher Foley <christopher.foley@mrc-bsu.cam.ac.uk> and James R Staley <js16174@bristol.ac.uk>
}
