% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyprcoloc.R
\name{ind.snp.score}
\alias{ind.snp.score}
\title{ind.snp.score}
\usage{
ind.snp.score(Q, snp.scores)
}
\arguments{
\item{Q}{the number of traits}

\item{snp.scores}{vector of per snp contributions to the posterior probability of colocalization}
}
\description{
ind.snp.score
}
