% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyprmtc.R
\name{print.hyprmtc}
\alias{print.hyprmtc}
\title{Print HyPrMTC}
\usage{
\method{print}{hyprmtc}(x, ...)
}
\arguments{
\item{x}{an object of class "hyprmtc"}
}
\description{
print method for class "hyprmtc"
}
\author{
Christopher Foley <christopher.foley@mrc-bsu.cam.ac.uk> and James R Staley <js16174@bristol.ac.uk>
}
