% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyprcoloc.R
\name{prior}
\alias{prior}
\title{prior}
\usage{
prior(p.1, gamma, k)
}
\arguments{
\item{p.1}{probability of one trait having a causal variant in the genetic region}

\item{gamma}{step probability}

\item{k}{number of traits}
}
\description{
prior
}
