% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyprcoloc.R
\name{rapid.hyprcoloc}
\alias{rapid.hyprcoloc}
\title{rapid.hyprcoloc}
\usage{
rapid.hyprcoloc(Zsq, Wsq, prior.1, prior.2, uniform.priors)
}
\arguments{
\item{Zsq}{matrix of Z-scores}

\item{Wsq}{ratio matrix of prior standard deviation and observed standard errors squared}

\item{prior.1}{prior probability of a SNP being associated with one trait}

\item{prior.2}{1 - prior probability of a SNP being associated with an additional trait given that the SNP is associated with at least 1 other trait}

\item{uniform.priors}{uniform priors}
}
\description{
rapid.hyprcoloc
}
