% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyprcoloc.R
\name{regional.ABF}
\alias{regional.ABF}
\title{regional.ABF}
\usage{
regional.ABF(
  Z,
  W,
  snps.clc,
  rho,
  trait.cor,
  sample.overlap,
  epsilon,
  reg.thresh,
  prior.1,
  prior.2,
  prior.3,
  prior.4,
  flag,
  test.2,
  reg.steps,
  cor.adj.priors,
  uniform.priors,
  branch.jump,
  Zsq,
  Wsq,
  ind.traits
)
}
\arguments{
\item{Z}{matrix of Z-scores}

\item{W}{ratio matrix of prior standard deviation and observed standard errors}

\item{snps.clc}{SNPs colocalisation}

\item{rho}{LD matrix}

\item{trait.cor}{correlation matrix between traits}

\item{sample.overlap}{matrix of sample overlap between traits}

\item{epsilon}{tolerance parameter}

\item{reg.thresh}{regional probability threshold}

\item{prior.1}{prior probability of a SNP being associated with one trait}

\item{prior.2}{1 - prior probability of a SNP being associated with an additional trait given that the SNP is associated with at least 1 other trait}

\item{prior.3}{prior probability that a trait contains a second causal variant given it contains one already}

\item{prior.4}{1 - prior probability that trait two co-localises with trait one given traits one and two already share a causal variant and trait one contains a second causal variant}

\item{flag}{flag variable}

\item{test.2}{test for 2CV}

\item{reg.steps}{regional step paramter}

\item{cor.adj.priors}{correlation adjusted priors}

\item{uniform.priors}{uniform priors}

\item{branch.jump}{branch jump}

\item{Zsq}{matrix of Z-scores squared}

\item{Wsq}{matrix of W squared}

\item{ind.traits}{are the traits independent or to be treated as independent}
}
\description{
regional.ABF
}
