% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyprcoloc.R
\name{samp.reduc}
\alias{samp.reduc}
\title{samp.reduc}
\usage{
samp.reduc(
  Z,
  W,
  ld.matrix,
  trait.cor,
  sample.overlap,
  trts,
  window.size = dim(Z)[1],
  sentinel = 0,
  Zsq,
  Wsq
)
}
\arguments{
\item{Z}{matrix of Z-scores}

\item{W}{ratio matrix of prior standard deviation and observed standard errors}

\item{ld.matrix}{LD matrix}

\item{trait.cor}{correlation matrix between traits}

\item{sample.overlap}{matrix of sample overlap between traits}

\item{trts}{vector of traits}

\item{window.size}{size of window for 2CV testing}

\item{sentinel}{sentinel variant}

\item{Zsq}{matrix of Z-scores squared}

\item{Wsq}{matrix of W squared}
}
\description{
samp.reduc
}
