% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyprcoloc.R
\name{sensitivity.plot}
\alias{sensitivity.plot}
\title{sensitivity.plot}
\usage{
sensitivity.plot(
  effect.est,
  effect.se,
  binary.outcomes = rep(0, dim(effect.est)[2]),
  trait.subset = c(1:dim(effect.est)[2]),
  trait.names = c(1:dim(effect.est)[2]),
  snp.id = c(1:dim(effect.est)[1]),
  ld.matrix = diag(1, dim(effect.est)[1], dim(effect.est)[1]),
  trait.cor = diag(1, dim(effect.est)[2], dim(effect.est)[2]),
  sample.overlap = matrix(rep(1, dim(effect.est)[2]^2), nrow = dim(effect.est)[2]),
  bb.alg = TRUE,
  bb.selection = "regional",
  reg.steps = 1,
  reg.thresh = c(0.6, 0.7, 0.8, 0.9),
  align.thresh = c(0.6, 0.7, 0.8, 0.9),
  prior.1 = 1e-04,
  prior.c = c(0.02, 0.01, 0.005),
  prior.12 = NULL,
  uniform.priors = FALSE,
  ind.traits = TRUE,
  equal.thresholds = FALSE,
  similarity.matrix = FALSE
)
}
\arguments{
\item{effect.est}{matrix of beta values}

\item{effect.se}{matrix of se values}

\item{binary.outcomes}{a binary vector depicting binary traits}

\item{trait.subset}{vector of traits from the full trait list for trageted coloclaisation analysis}

\item{trait.names}{vector of trait names corresponding to the columns in the effect.est matrix}

\item{snp.id}{vector of SNP IDs}

\item{ld.matrix}{LD matrix}

\item{trait.cor}{correlation matrix between traits}

\item{sample.overlap}{matrix of sample overlap between traits}

\item{bb.alg}{branch and bound algorithm}

\item{bb.selection}{branch and bound algorithm type}

\item{reg.steps}{regional step paramter}

\item{align.thresh}{a vector of alignment probability thresholds}

\item{prior.1}{prior probability of a SNP being associated with one trait}

\item{prior.c}{vector of conditional colocalization priors: where prior.c is the probability of a SNP being associated with an additional trait given that the SNP is associated with at least 1 other trait}

\item{prior.12}{COLOC prior p12: prior probability of a SNP being associated with any two traits}

\item{ind.traits}{are the traits independent or to be treated as independent}

\item{equal.thresholds}{fix the regional and alignment thresholds to be equal}

\item{n.cvs}{number of causal variants}

\item{window.size}{size of window for 2CV testing}

\item{sentinel}{sentinel variant}

\item{epsilon}{tolerance parameter}

\item{reg.thres}{a vector of regional probability thresholds}

\item{reg.tol}{regional tolerance parameter}

\item{prior.3}{prior probability that a trait contains a second causal variant given it contains one already}

\item{prior.4}{1 - prior probability that trait two co-localises with trait one given traits one and two already share a causal variant and trait one contains a second causal variant}

\item{unifrom.priors}{uniform priors}
}
\description{
sensitivity.plot is a function which repeatedly calls the hyprcoloc function to compute a similarity matrix which illustrates how strongly clustered/colocalized pairs of traits are across different input thresholds and priors
}
