% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyprcoloc.R
\name{cred.sets}
\alias{cred.sets}
\title{cred.sets}
\usage{
cred.sets(res, value = 0.95)
}
\arguments{
\item{res}{list of results from hyprcoloc when setting "snpscores = TRUE"}

\item{value}{the sum of the probabilities of the snps included in the credible set}
}
\description{
cred.sets
}
