% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyprcoloc.R
\name{hyprcoloc}
\alias{hyprcoloc}
\title{HyPrColoc}
\usage{
hyprcoloc(
  effect.est,
  effect.se,
  binary.outcomes = rep(0, dim(effect.est)[2]),
  trait.subset = c(1:dim(effect.est)[2]),
  trait.names = c(1:dim(effect.est)[2]),
  snp.id = c(1:dim(effect.est)[1]),
  ld.matrix = diag(1, dim(effect.est)[1], dim(effect.est)[1]),
  trait.cor = diag(1, dim(effect.est)[2], dim(effect.est)[2]),
  sample.overlap = matrix(rep(1, dim(effect.est)[2]^2), nrow = dim(effect.est)[2]),
  bb.alg = TRUE,
  bb.selection = "regional",
  reg.steps = 1,
  reg.thresh = "default",
  align.thresh = "default",
  prior.1 = 1e-04,
  prior.c = 0.02,
  prior.12 = NULL,
  sensitivity = FALSE,
  sense.1 = 1,
  sense.2 = 2,
  uniform.priors = FALSE,
  ind.traits = FALSE,
  snpscores = FALSE
)
}
\arguments{
\item{effect.est}{matrix of snp regression coefficients (i.e. regression beta values) in the genomic region}

\item{effect.se}{matrix of standard errors associated with the beta values}

\item{binary.outcomes}{a binary vector of dimension the number of traits: 1 represents a binary trait 0 otherwise}

\item{trait.subset}{vector of trait names (or number) from the full trait list: used for trageted colocalization analysis in a region}

\item{trait.names}{vector of trait names corresponding to the columns in the effect.est matrix}

\item{snp.id}{vector of SNP IDs}

\item{ld.matrix}{LD matrix}

\item{trait.cor}{matrix of pairwise correlations between traits}

\item{sample.overlap}{matrix of pairwise sample overlap between traits}

\item{bb.alg}{branch and bound algorithm: TRUE, employ BB algorithm; FALSE, do not}

\item{bb.selection}{branch and bound algorithm type, e.g. regional or alignment selection}

\item{reg.steps}{regional step paramter}

\item{reg.thresh}{threshold probability beyond which traits are believed to share a regional association signal}

\item{align.thresh}{threshold probability beyond which traits are believed to align at a single causal variant}

\item{prior.1}{prior probability of a SNP being associated with one trait}

\item{prior.c}{conditional colocalization prior: probability of a SNP being associated with an additional trait given that the SNP is associated with at least 1 other trait}

\item{prior.12}{COLOC prior p12: prior probability of a SNP being associated with any two traits}

\item{sensitivity}{perform senstivity analysis}

\item{sense.1}{first sensitivity analysis}

\item{sense.2}{second sensitivity analysis}

\item{uniform.priors}{uniform priors}

\item{ind.traits}{are the traits independent or to be treated as independent}

\item{snpscores}{output estimated posterior probability explained each SNP}
}
\value{
A data.frame of HyPrColoc results: each row is a cluster of colocalized traits or is coded NA (if no colocalization is identified)

If snpscores = TRUE: additionally returns a list of posterior probability explained by each SNPs and for each cluster of colocalized traits identified
}
\description{
\code{hyprcoloc} is a function used to identify clusters of colocalized traits and candidate causal SNPs in genomic regions
}
\examples{
# Regression coefficients and standard errors from ten GWAS studies
# (Traits 1-5, 6-8 & 9-10 are the clusters of colocalized traits)
betas <- hyprcoloc::test.betas
head(betas)
ses <- hyprcoloc::test.ses
head(ses)

# Trait names and SNP IDs
traits <- paste0("T", 1:10)
rsid <- rownames(betas)

# Colocalisation analyses
results <- hyprcoloc(betas, ses, trait.names = traits, snp.id = rsid)
}
\author{
Christopher Foley \href{mailto:chris.neal.foley@gmail.com}{chris.neal.foley@gmail.com} & James Staley \href{mailto:jrstaley95@gmail.com}{jrstaley95@gmail.com}
}
