% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyprcoloc.R
\name{print.hyprcoloc}
\alias{print.hyprcoloc}
\title{Print HyPrColoc}
\usage{
print.hyprcoloc(x, ...)
}
\arguments{
\item{x}{an object of class "hyprcoloc"}

\item{...}{Other arguments passed to or from other methods}
}
\description{
print method for class "hyprcoloc"
}
\author{
Christopher N Foley (University of Cambridge) <chris.neal.foley@gmail.com> and James R Staley (University of Bristol) <jrstaley95@gmail.com>
}
