% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyprcoloc.R
\name{print.hyprcoloc}
\alias{print.hyprcoloc}
\title{Print HyPrColoc}
\usage{
\method{print}{hyprcoloc}(x, ...)
}
\arguments{
\item{x}{an object of class "hyprcoloc"}
}
\description{
print method for class "hyprcoloc"
}
\author{
Christopher Foley <christopher.foley@mrc-bsu.cam.ac.uk> and James R Staley <james.staley@bristol.ac.uk>
}
