% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyprcoloc.R
\name{print.hyprcoloc}
\alias{print.hyprcoloc}
\title{Print hyprcoloc}
\usage{
\method{print}{hyprcoloc}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"hyprcoloc"}}
}
\description{
print method for class \code{"hyprcoloc"}.
}
\author{
Christopher Foley \href{mailto:chris.neal.foley@gmail.com}{chris.neal.foley@gmail.com} & James Staley \href{mailto:jrstaley95@gmail.com}{jrstaley95@gmail.com}
}
