% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyprcoloc.R
\name{sensitivity.plot}
\alias{sensitivity.plot}
\title{Sensitivity plot}
\usage{
sensitivity.plot(
  effect.est,
  effect.se,
  binary.outcomes = rep(0, dim(effect.est)[2]),
  trait.subset = c(1:dim(effect.est)[2]),
  trait.names = c(1:dim(effect.est)[2]),
  snp.id = c(1:dim(effect.est)[1]),
  ld.matrix = diag(1, dim(effect.est)[1], dim(effect.est)[1]),
  trait.cor = diag(1, dim(effect.est)[2], dim(effect.est)[2]),
  sample.overlap = matrix(rep(1, dim(effect.est)[2]^2), nrow = dim(effect.est)[2]),
  bb.alg = TRUE,
  bb.selection = "regional",
  reg.steps = 1,
  reg.thresh = c(0.6, 0.7, 0.8, 0.9),
  align.thresh = c(0.6, 0.7, 0.8, 0.9),
  prior.1 = 1e-04,
  prior.c = c(0.02, 0.01, 0.005),
  prior.12 = NULL,
  uniform.priors = FALSE,
  ind.traits = TRUE,
  equal.thresholds = FALSE,
  similarity.matrix = FALSE
)
}
\arguments{
\item{effect.est}{matrix of snp regression coefficients (i.e. regression beta values) in the genomic region}

\item{effect.se}{matrix of standard errors associated with the beta values}

\item{binary.outcomes}{a binary vector of dimension the number of traits: 1 represents a binary trait 0 otherwise}

\item{trait.subset}{vector of trait names (or number) from the full trait list: used for trageted colocalization analysis in a region}

\item{trait.names}{vector of trait names corresponding to the columns in the effect.est matrix}

\item{snp.id}{vector of SNP IDs}

\item{ld.matrix}{LD matrix}

\item{trait.cor}{matrix of pairwise correlations between traits}

\item{sample.overlap}{matrix of pairwise sample overlap between traits}

\item{bb.alg}{branch and bound algorithm: TRUE, employ BB algorithm; FALSE, do not}

\item{bb.selection}{branch and bound algorithm type, e.g. regional or alignment selection}

\item{reg.steps}{regional step paramter}

\item{reg.thresh}{a vector of regional probability thresholds}

\item{align.thresh}{a vector of alignment probability thresholds}

\item{prior.1}{prior probability of a SNP being associated with one trait}

\item{prior.c}{vector of conditional colocalization priors: where prior.c is the probability of a SNP being associated with an additional trait given that the SNP is associated with at least 1 other trait}

\item{prior.12}{COLOC prior p12: prior probability of a SNP being associated with any two traits}

\item{uniform.priors}{uniform priors}

\item{ind.traits}{are the traits independent or to be treated as independent}

\item{equal.thresholds}{fix the regional and alignment thresholds to be equal}

\item{similarity.matrix}{To be viewed as a similarity matrix. Default FALSE.}
}
\description{
\code{sensitivity.plot} is a function which repeatedly calls the hyprcoloc function to compute a similarity matrix which illustrates how strongly clustered/colocalized pairs of traits are across different input thresholds and priors
}
\author{
Christopher Foley \href{mailto:chris.neal.foley@gmail.com}{chris.neal.foley@gmail.com} & James Staley \href{mailto:jrstaley95@gmail.com}{jrstaley95@gmail.com}
}
